requirements.txt
matplotlib
numpy
pillow
tensorflow
watchdog